Pronto
Tengo un pero.
Siento que resolví un problema por instinto o casualidad, pero al mismo tiempo sé que va ahí.
Te cuento más tarde! 😉🔍
