+++ frontend/src/Pagina/PerfilDeParticipante/ListaTitulos.jsx
@@ -1,81 +1,8 @@
-import { useEffect, useState } from "react";
-import { deleteTitulos, getTitulos, postTitulos } from "../../API/Api";
-
-/*
-
-Que muestra mi pantalla?
- - Input para ingresar titulo
- - Lista que muestra los titulo
- - Boton que guarda el titulo
- - Botones de borrar
-
-Variables del estado:
- - Estado para el input
- - Estado para la lista de titulos
-
-Acciones de la pantalla:
- - Borrar un titulo
- - Crear un titulo
- - Mostrar todos los titulos
-
-
- Pasos:
- 1. declarar los estados
- 2. asociar los estado al HTML
- 3. hacer una funcion para cada accion
-
-
-
- Agregar mensaje de error:
- 1. Agregar nuevo estado booleando con valor por defecto false
- 2. En el HTML, cuando el estado tengo valor igual a true, mostrar un mensaje de error
- 3. Capturar posibles errores y poner el estado en true cuando hay error
-
-  Usar try-catch para capturar errores, buscar en interne como funciona.
-*/
+import useTitulos from "../../Hooks/useTitulos";
 
 const ListaTitulos = ({ idCowboy }) => {
-  // Estados
-  const [input, setInput] = useState("");
-  const [titulos, setTitulos] = useState([]);
-  const [error, setError] = useState(false);
-
-  // Hooks
-  useEffect(() => {
-    mostrarTitulo();
-  }, []);
-
-  // Acciones
-  const mostrarTitulo = async () => {
-    // llamar a la api de obtener
-    const lista = await getTitulos(idCowboy);
-    setTitulos(lista);
-  };
-
-  const crearTitulo = async () => {
-    // llamar a la api de crear
-    const titulo = {
-      titulo: input,
-    };
-    try {
-      const tituloCreado = await postTitulos(idCowboy, titulo);
-      const nuevaLista = [...titulos, tituloCreado];
-      setTitulos(nuevaLista);
-      setError(false)
-    } catch (err) {
-      setError(true);
-      console.log("Hubo un error!!!!!");
-    }
-    setInput("");
-  };
-
-  const borrarTitulo = async (id) => {
-    // llamar a la api de borrar
-    await deleteTitulos(id);
-
-    const listaBorrada = titulos.filter((titulo) => titulo.id != id);
-    setTitulos(listaBorrada);
-  };
+  // Utilizamos el hook useTitulos
+  const { borrarTitulo, crearTitulo, error, titulos, setInput, input } = useTitulos(idCowboy);
 
   return (
     <>
